Extract generic concept
