+++ src/main/java/org/testng/internal/ExpectedExceptionsHolder.java
@@ -10,40 +10,12 @@ import java.util.regex.Pattern;
  * A class that contains the expected exceptions and the message regular expression.
  * @author cbeust
  */
-public class ExpectedExceptionsHolder {
-  private final Class<?>[] expectedClasses;
+public class ExpectedExceptionsHolder extends AbstractExpectedExceptionsHolder {
   private final String messageRegExp;
 
   public ExpectedExceptionsHolder(Class<?>[] expectedClasses, String messageRegExp) {
-    this.expectedClasses = expectedClasses;
+    super(expectedClasses);
     this.messageRegExp = messageRegExp;
-  }
-
-  /**
-   * @param ite The exception that was just thrown
-   * @return true if the exception that was just thrown is part of the
-   * expected exceptions
-   */
-  public boolean isExpectedException(Throwable ite) {
-    if (expectedClasses == null) {
-      return false;
-    }
-
-    // TestException is the wrapper exception that TestNG will be throwing when an exception was
-    // expected but not thrown
-    if (ite.getClass() == TestException.class) {
-      return false;
-    }
-
-    Class<?> realExceptionClass= ite.getClass();
-
-    for (Class<?> exception : expectedClasses) {
-      if (exception.isAssignableFrom(realExceptionClass) && messageRegExpMatches(ite)) {
-        return true;
-      }
-    }
-
-    return false;
   }
 
   /**
@@ -51,7 +23,8 @@ public class ExpectedExceptionsHolder {
    *   null             true    false
    *   non-null         true    match
    */
-  private boolean messageRegExpMatches(Throwable ite) {
+  @Override
+  protected boolean isExceptionMatches(Throwable ite) {
     if (".*".equals(messageRegExp)) {
       return true;
     } else {
@@ -60,35 +33,9 @@ public class ExpectedExceptionsHolder {
     }
   }
 
-  public TestException wrongException(Throwable ite) {
-    if (messageRegExpMatches(ite)) {
-      return new TestException("Expected exception of " +
-                               getExpectedExceptionsPluralize()
-                               + " but got " + ite, ite);
-    } else {
-      return new TestException("The exception was thrown with the wrong message:" +
-                               " expected \"" + messageRegExp + "\"" +
-                               " but got \"" + ite.getMessage() + "\"", ite);
-    }
-  }
-
-  public TestException noException(ITestNGMethod testMethod) {
-    if (expectedClasses == null || expectedClasses.length == 0) {
-      return null;
-    }
-    return new TestException("Method " + testMethod + " should have thrown an exception of "
-                             + getExpectedExceptionsPluralize());
-  }
-
-  private String getExpectedExceptionsPluralize() {
-    StringBuilder sb = new StringBuilder();
-    if (expectedClasses.length > 1) {
-      sb.append("any of types ");
-      sb.append(Arrays.toString(expectedClasses));
-    } else {
-      sb.append("type ");
-      sb.append(expectedClasses[0]);
-    }
-    return sb.toString();
+  protected String getWrongExceptionMessage(Throwable ite) {
+    return "The exception was thrown with the wrong message:" +
+           " expected \"" + messageRegExp + "\"" +
+           " but got \"" + ite.getMessage() + "\"";
   }
 }
